refactor(web): split request kind filter surfaces
Extract the shared request-type filter into a focused component.

Move request-kind styles into dedicated files to restore line-budget checks.
