+++ web/src/components/AdminRecentRequestsPanel.tsx
@@ -6,18 +6,13 @@ import type { AdminTranslations } from '../i18n'
 import { cleanedRequestLogBodySummary } from '../requestLogBodySummary'
 import { Icon } from '../lib/icons'
 import {
-  buildRequestKindQuickFilterSelection,
-  buildVisibleRequestKindOptions,
-  hasActiveRequestKindQuickFilters,
-  mergeRequestKindCatalog,
-  summarizeRequestKindQuickFilters,
-  summarizeSelectedRequestKinds,
   type TokenLogRequestKindOption,
   type TokenLogRequestKindQuickBilling,
   type TokenLogRequestKindQuickProtocol,
 } from '../tokenLogRequestKinds'
 import RequestKindBadge from './RequestKindBadge'
 import RequestIpDiagnostics from './RequestIpDiagnostics'
+import AdminRecentRequestsRequestKindFilter from './AdminRecentRequestsRequestKindFilter'
 import AdminLoadingRegion from './AdminLoadingRegion'
 import AdminTablePagination from './AdminTablePagination'
 import AdminTableShell from './AdminTableShell'
@@ -38,19 +33,6 @@ import SearchableFacetSelect from './SearchableFacetSelect'
 import { StatusBadge, type StatusTone } from './StatusBadge'
 import { Button } from './ui/button'
 import {
-  Drawer,
-  DrawerContent,
-  DrawerDescription,
-  DrawerHeader,
-  DrawerTitle,
-} from './ui/drawer'
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-} from './ui/dropdown-menu'
-import {
   Select,
   SelectContent,
   SelectGroup,
@@ -59,7 +41,6 @@ import {
   SelectSeparator,
   SelectTrigger,
 } from './ui/select'
-import SegmentedTabs from './ui/SegmentedTabs'
 import { TableBody, TableCell, TableHead, TableHeader, TableRow } from './ui/table'
 import { Tooltip, TooltipContent, TooltipTrigger } from './ui/tooltip'
 import { useViewportMode } from '../lib/responsive'
@@ -122,27 +103,8 @@ type LogBodiesLoadState =
   | { status: 'loading' }
   | { status: 'ready'; value: RequestLogBodies }
   | { status: 'error'; message: string }
-const requestKindBillingQuickFilterOptions = [
-  { value: 'all', label: 'Any' },
-  { value: 'billable', label: 'Paid' },
-  { value: 'non_billable', label: 'Free' },
-] as const
-const requestKindProtocolQuickFilterOptions = [
-  { value: 'all', label: 'Any' },
-  { value: 'mcp', label: 'MCP' },
-  { value: 'api', label: 'API' },
-] as const
 const recentRequestsAllFilterValue = '__all__'
 const recentRequestsCompactAllLabel = 'All'
-
-function resolveRequestKindProtocolGroup(
-  option: TokenLogRequestKindOption,
-): 'api' | 'mcp' {
-  if (option.protocol_group === 'api' || option.key.startsWith('api:')) {
-    return 'api'
-  }
-  return 'mcp'
-}
 function statusTone(status: string): StatusTone {
   const normalized = status.trim().toLowerCase()
   if (normalized === 'success') return 'success'
@@ -584,19 +546,6 @@ function renderOutcomeFacetLabel(
           : selectionEffectLabel(value, strings)
   return <StatusBadge tone={tone}>{label}</StatusBadge>
 }
-function summarizeRequestKindTrigger(
-  effectiveSelectedRequestKinds: string[],
-  hasActiveQuickRequestKindFilters: boolean,
-  requestKindQuickSummary: string,
-  requestKindSummary: string,
-  language: Language,
-  allLabel: string,
-): string {
-  if (hasActiveQuickRequestKindFilters) return requestKindQuickSummary
-  if (effectiveSelectedRequestKinds.length === 0) return allLabel
-  if (effectiveSelectedRequestKinds.length <= 2) return requestKindSummary
-  return language === 'zh' ? `已选 ${effectiveSelectedRequestKinds.length} 项` : `${effectiveSelectedRequestKinds.length} selected`
-}
 function RecentRequestDetails({
   log,
   logBodiesState,
@@ -837,7 +786,6 @@ export default function AdminRecentRequestsPanel({
   const [expandedLogs, setExpandedLogs] = useState<Set<number>>(() => new Set())
   const [logBodiesById, setLogBodiesById] = useState<Record<number, LogBodiesLoadState>>({})
   const [headerFiltersTarget, setHeaderFiltersTarget] = useState<HTMLElement | null>(null)
-  const [requestKindFilterOpen, setRequestKindFilterOpen] = useState(false)
   const logBodyControllersRef = useRef<Map<number, AbortController>>(new Map())
   const viewportMode = useViewportMode()
   const isSmallViewport = viewportMode === 'small'
@@ -919,85 +867,6 @@ export default function AdminRecentRequestsPanel({
     },
     [triggerLoadLogBodies],
   )
-  const normalizedSelectedRequestKinds = useMemo(
-    () => Array.from(new Set(selectedRequestKinds.map((value) => value.trim()).filter(Boolean))),
-    [selectedRequestKinds],
-  )
-  const requestKindCatalog = useMemo(
-    () => mergeRequestKindCatalog(requestKindOptions),
-    [requestKindOptions],
-  )
-  const requestKindQuickFilters = useMemo(
-    () => ({
-      billing: requestKindQuickBilling,
-      protocol: requestKindQuickProtocol,
-    }),
-    [requestKindQuickBilling, requestKindQuickProtocol],
-  )
-  const hasActiveQuickRequestKindFilters = useMemo(
-    () => hasActiveRequestKindQuickFilters(requestKindQuickFilters),
-    [requestKindQuickFilters],
-  )
-  const quickSelection = useMemo(
-    () => buildRequestKindQuickFilterSelection(requestKindOptions, requestKindQuickFilters),
-    [requestKindOptions, requestKindQuickFilters],
-  )
-  const effectiveSelectedRequestKinds = useMemo(
-    () => (hasActiveQuickRequestKindFilters ? quickSelection : normalizedSelectedRequestKinds),
-    [hasActiveQuickRequestKindFilters, normalizedSelectedRequestKinds, quickSelection],
-  )
-  const visibleRequestKindOptions = useMemo(
-    () =>
-      buildVisibleRequestKindOptions(
-        effectiveSelectedRequestKinds,
-        requestKindCatalog,
-        Object.fromEntries(requestKindCatalog.map((option) => [option.key, option])),
-      ),
-    [effectiveSelectedRequestKinds, requestKindCatalog],
-  )
-  const requestKindSummary = useMemo(
-    () =>
-      summarizeSelectedRequestKinds(
-        effectiveSelectedRequestKinds,
-        visibleRequestKindOptions,
-        strings.logs.filters.requestTypeAll,
-      ),
-    [effectiveSelectedRequestKinds, strings.logs.filters.requestTypeAll, visibleRequestKindOptions],
-  )
-  const requestKindQuickSummary = useMemo(
-    () => summarizeRequestKindQuickFilters(requestKindQuickFilters),
-    [requestKindQuickFilters],
-  )
-  const requestKindClearDisabled =
-    effectiveSelectedRequestKinds.length === 0 && !hasActiveQuickRequestKindFilters
-  const requestKindColumnGroups = useMemo(() => {
-    const api: TokenLogRequestKindOption[] = []
-    const mcp: TokenLogRequestKindOption[] = []
-    for (const option of visibleRequestKindOptions) {
-      if (resolveRequestKindProtocolGroup(option) === 'api') {
-        api.push(option)
-      } else {
-        mcp.push(option)
-      }
-    }
-    return { api, mcp }
-  }, [visibleRequestKindOptions])
-  const requestKindTriggerSummary = useMemo(() => {
-    return summarizeRequestKindTrigger(
-      effectiveSelectedRequestKinds,
-      hasActiveQuickRequestKindFilters,
-      requestKindQuickSummary,
-      requestKindSummary,
-      language,
-      recentRequestsCompactAllLabel,
-    )
-  }, [
-    effectiveSelectedRequestKinds.length,
-    hasActiveQuickRequestKindFilters,
-    language,
-    requestKindQuickSummary,
-    requestKindSummary,
-  ])
   const outcomeValue = outcomeFilter
     ? `${outcomeFilter.kind}:${outcomeFilter.value}`
     : recentRequestsAllFilterValue
@@ -1026,187 +895,20 @@ export default function AdminRecentRequestsPanel({
       ? 'user-console-mobile-kv user-console-mobile-kv--stacked'
       : 'admin-mobile-kv admin-mobile-kv--stacked'
   const headerCopyVisible = showHeaderCopy && (title.trim().length > 0 || description.trim().length > 0)
-  const handleClearRequestKinds = useCallback(() => {
-    if (requestKindClearDisabled) return
-    onClearRequestKinds()
-  }, [onClearRequestKinds, requestKindClearDisabled])
-  const renderRequestKindOptionsList = useCallback(
-    (
-      options: TokenLogRequestKindOption[],
-      groupLabel: string,
-      container: 'dropdown' | 'drawer',
-    ) => (
-      <div className="token-request-kind-group">
-        <div className="token-request-kind-group-label">{groupLabel}</div>
-        {options.length === 0 ? (
-          <div className="token-request-kind-empty">{strings.logs.filters.requestTypeEmpty}</div>
-        ) : (
-          <div className="token-request-kind-group-options">
-            {options.map((option) => {
-              const checked = effectiveSelectedRequestKinds.includes(option.key)
-              const content = (
-                <span className="recent-requests-request-kind-option">
-                  <RequestKindBadge requestKindKey={option.key} requestKindLabel={option.label} size="sm" />
-                  <span className="recent-requests-request-kind-count">
-                    {`x${option.count ?? 0}`}
-                  </span>
-                </span>
-              )
-              if (container === 'drawer') {
-                return (
-                  <button
-                    key={option.key}
-                    type="button"
-                    role="checkbox"
-                    aria-checked={checked}
-                    className={`recent-requests-request-kind-drawer-item${
-                      checked ? ' recent-requests-request-kind-drawer-item--checked' : ''
-                    }`}
-                    onClick={() => onToggleRequestKind(option.key)}
-                  >
-                    <span className="recent-requests-request-kind-drawer-mark" aria-hidden="true">
-                      {checked ? (
-                        <Icon icon="mdi:check" width={16} height={16} />
-                      ) : null}
-                    </span>
-                    {content}
-                  </button>
-                )
-              }
-              return (
-                <DropdownMenuCheckboxItem
-                  key={option.key}
-                  className="cursor-pointer recent-requests-request-kind-item"
-                  checked={checked}
-                  onSelect={(event) => event.preventDefault()}
-                  onCheckedChange={() => onToggleRequestKind(option.key)}
-                >
-                  {content}
-                </DropdownMenuCheckboxItem>
-              )
-            })}
-          </div>
-        )}
-      </div>
-    ),
-    [effectiveSelectedRequestKinds, onToggleRequestKind, strings.logs.filters.requestTypeEmpty],
-  )
-  const renderRequestKindFiltersContent = useCallback(
-    (container: 'dropdown' | 'drawer') => (
-      <div
-        className={[
-          'token-request-kind-panel',
-          `token-request-kind-panel--${container}`,
-        ].join(' ')}
-      >
-        <div className="token-request-kind-panel-header">
-          <div className="token-request-kind-panel-title">{strings.logs.filters.requestType}</div>
-          <Button
-            type="button"
-            variant="ghost"
-            size="xs"
-            className="token-request-kind-clear"
-            disabled={requestKindClearDisabled}
-            onClick={handleClearRequestKinds}
-          >
-            {strings.users.clear}
-          </Button>
-        </div>
-        <div className="token-request-kind-layout">
-          <div className="token-request-kind-quick-filters">
-            <div className="token-request-kind-quick-cell">
-              <div className="token-request-kind-group-label">{strings.logs.filters.billingGroup}</div>
-              <SegmentedTabs<TokenLogRequestKindQuickBilling>
-                value={requestKindQuickBilling}
-                onChange={(next) => onRequestKindQuickFiltersChange(next, requestKindQuickProtocol)}
-                options={requestKindBillingQuickFilterOptions}
-                ariaLabel={strings.logs.filters.billingGroup}
-                className="token-request-quick-segmented"
-                smallViewportBehavior={container === 'drawer' ? 'buttons' : 'select'}
-              />
-            </div>
-            <div className="token-request-kind-quick-cell">
-              <div className="token-request-kind-group-label">{strings.logs.filters.protocolGroup}</div>
-              <SegmentedTabs<TokenLogRequestKindQuickProtocol>
-                value={requestKindQuickProtocol}
-                onChange={(next) => onRequestKindQuickFiltersChange(requestKindQuickBilling, next)}
-                options={requestKindProtocolQuickFilterOptions}
-                ariaLabel={strings.logs.filters.protocolGroup}
-                className="token-request-quick-segmented"
-                smallViewportBehavior={container === 'drawer' ? 'buttons' : 'select'}
-              />
-            </div>
-          </div>
-          <div className="token-request-kind-columns">
-            {renderRequestKindOptionsList(requestKindColumnGroups.api, 'API', container)}
-            {renderRequestKindOptionsList(requestKindColumnGroups.mcp, 'MCP', container)}
-          </div>
-        </div>
-      </div>
-    ),
-    [
-      handleClearRequestKinds,
-      onRequestKindQuickFiltersChange,
-      requestKindClearDisabled,
-      requestKindColumnGroups.api,
-      requestKindColumnGroups.mcp,
-      requestKindQuickBilling,
-      requestKindQuickProtocol,
-      renderRequestKindOptionsList,
-      strings.logs.filters.billingGroup,
-      strings.logs.filters.protocolGroup,
-      strings.logs.filters.requestType,
-      strings.users.clear,
-    ],
-  )
   const renderFilters = (className?: string) => (
     <div className={['panel-actions recent-requests-filters', className].filter(Boolean).join(' ')}>
-      <div className="recent-requests-filter-field recent-requests-filter-field--request-kind">
-        <span className="recent-requests-filter-label">{strings.logs.filters.requestType}</span>
-        {isSmallViewport ? (
-          <Drawer
-            open={requestKindFilterOpen}
-            onOpenChange={setRequestKindFilterOpen}
-            shouldScaleBackground={false}
-          >
-            <button
-              type="button"
-              className="recent-requests-filter-select-trigger recent-requests-filter-select-trigger--menu"
-              aria-label={`${strings.logs.filters.requestType}: ${requestKindTriggerSummary}`}
-              onClick={() => setRequestKindFilterOpen(true)}
-            >
-              <span className="recent-requests-filter-select-text">{requestKindTriggerSummary}</span>
-              <Icon icon="mdi:chevron-down" width={16} height={16} aria-hidden="true" />
-            </button>
-            <DrawerContent className="token-request-kind-drawer">
-              <DrawerHeader className="sr-only">
-                <DrawerTitle>{strings.logs.filters.requestType}</DrawerTitle>
-                <DrawerDescription>{strings.logs.descriptionFallback}</DrawerDescription>
-              </DrawerHeader>
-              {renderRequestKindFiltersContent('drawer')}
-            </DrawerContent>
-          </Drawer>
-        ) : (
-          <DropdownMenu open={requestKindFilterOpen} onOpenChange={setRequestKindFilterOpen}>
-            <DropdownMenuTrigger asChild>
-              <button
-                type="button"
-                className="recent-requests-filter-select-trigger recent-requests-filter-select-trigger--menu"
-                aria-label={`${strings.logs.filters.requestType}: ${requestKindTriggerSummary}`}
-              >
-                <span className="recent-requests-filter-select-text">{requestKindTriggerSummary}</span>
-                <Icon icon="mdi:chevron-down" width={16} height={16} aria-hidden="true" />
-              </button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent
-              align="end"
-              className="token-request-kind-menu recent-requests-filter-menu recent-requests-filter-menu--request-kind"
-            >
-              {renderRequestKindFiltersContent('dropdown')}
-            </DropdownMenuContent>
-          </DropdownMenu>
-        )}
-      </div>
+      <AdminRecentRequestsRequestKindFilter
+        language={language}
+        isSmallViewport={isSmallViewport}
+        strings={strings}
+        requestKindOptions={requestKindOptions}
+        requestKindQuickBilling={requestKindQuickBilling}
+        requestKindQuickProtocol={requestKindQuickProtocol}
+        selectedRequestKinds={selectedRequestKinds}
+        onRequestKindQuickFiltersChange={onRequestKindQuickFiltersChange}
+        onToggleRequestKind={onToggleRequestKind}
+        onClearRequestKinds={onClearRequestKinds}
+      />
       <div className="recent-requests-filter-field">
         <span className="recent-requests-filter-label">{strings.logs.filters.resultOrEffect}</span>
         <Select
