+++ web/src/styles/admin-recent-requests-request-kind-filter.css
@@ -1,0 +1,223 @@
+.token-request-kind-trigger {
+  min-width: min(100%, 248px);
+  justify-content: flex-start;
+  gap: 10px;
+  padding-inline: 12px;
+  min-height: 46px;
+}
+
+.token-request-kind-trigger-content {
+  display: flex;
+  flex: 1 1 auto;
+  min-width: 0;
+  align-items: flex-start;
+}
+
+.token-request-kind-trigger-summary {
+  max-width: 100%;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  font-weight: 700;
+}
+
+.token-request-kind-count {
+  display: inline-flex;
+  min-width: 1.9rem;
+  justify-content: center;
+  align-items: center;
+  padding: 0 7px;
+  border-radius: 999px;
+  background: hsl(var(--muted) / 0.68);
+  color: hsl(var(--foreground) / 0.82);
+  font-size: 0.76rem;
+  font-weight: 700;
+  line-height: 1.7;
+}
+
+.token-request-quick-segmented.segmented-tabs {
+  display: flex;
+  width: 100%;
+}
+
+.token-request-quick-segmented.segmented-tabs .segmented-tab {
+  flex: 1 1 0;
+  justify-content: center;
+  padding-inline: 12px;
+}
+
+.token-request-kind-menu {
+  width: min(46rem, calc(100vw - 1rem));
+  max-height: min(
+    var(--radix-dropdown-menu-content-available-height, calc(100dvh - 1rem)),
+    calc(100dvh - 1rem)
+  );
+  overflow-y: auto;
+  padding: 0;
+}
+
+.token-request-kind-menu [role='menuitemcheckbox'] {
+  align-items: flex-start;
+}
+
+.token-request-kind-panel {
+  display: grid;
+  gap: 14px;
+  padding: 14px;
+}
+
+.token-request-kind-panel-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+}
+
+.token-request-kind-panel-title {
+  min-width: 0;
+  font-size: 0.95rem;
+  font-weight: 800;
+  color: hsl(var(--foreground));
+}
+
+.token-request-kind-clear {
+  flex: 0 0 auto;
+  min-height: 2rem;
+  padding-inline: 0.2rem;
+  border-radius: 10px;
+  font-size: 0.76rem;
+  font-weight: 800;
+  letter-spacing: 0.02em;
+}
+
+.token-request-kind-layout {
+  display: grid;
+  gap: 14px;
+}
+
+.token-request-kind-quick-filters {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 12px;
+}
+
+.token-request-kind-quick-cell {
+  display: grid;
+  gap: 8px;
+  min-width: 0;
+}
+
+.token-request-kind-group-label {
+  font-size: 0.72rem;
+  font-weight: 700;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+  color: hsl(var(--muted-foreground));
+}
+
+.token-request-kind-columns {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 12px;
+  min-width: 0;
+  align-items: start;
+}
+
+.token-request-kind-group {
+  display: grid;
+  gap: 8px;
+  min-width: 0;
+  align-content: start;
+}
+
+.token-request-kind-group-options {
+  display: grid;
+  gap: 6px;
+}
+
+.token-request-kind-empty {
+  border: 1px dashed hsl(var(--border) / 0.76);
+  border-radius: 14px;
+  padding: 0.9rem 1rem;
+  color: hsl(var(--muted-foreground));
+  font-size: 0.84rem;
+  line-height: 1.35;
+}
+
+.recent-requests-request-kind-item {
+  width: 100%;
+  max-width: 100%;
+  padding-inline: 0.55rem;
+}
+
+.recent-requests-request-kind-drawer-item {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.75rem;
+  width: 100%;
+  padding: 0.78rem 0.8rem;
+  border: 1px solid hsl(var(--border) / 0.76);
+  border-radius: 16px;
+  background: hsl(var(--card) / 0.92);
+  color: hsl(var(--foreground));
+  text-align: left;
+  transition:
+    border-color 0.18s ease,
+    background-color 0.18s ease,
+    box-shadow 0.18s ease;
+}
+
+.recent-requests-request-kind-drawer-item:hover,
+.recent-requests-request-kind-drawer-item:focus-visible {
+  border-color: hsl(var(--ring) / 0.42);
+  background: hsl(var(--accent) / 0.54);
+  outline: none;
+}
+
+.recent-requests-request-kind-drawer-item--checked {
+  border-color: hsl(var(--primary) / 0.34);
+  background: hsl(var(--primary) / 0.1);
+}
+
+.recent-requests-request-kind-drawer-mark {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: 1.1rem;
+  min-width: 1.1rem;
+  height: 1.1rem;
+  margin-top: 0.18rem;
+  border-radius: 999px;
+  border: 1px solid hsl(var(--border) / 0.9);
+  color: hsl(var(--primary));
+}
+
+.token-request-kind-drawer {
+  max-height: min(88dvh, 46rem);
+  overflow-y: auto;
+}
+
+.token-request-kind-panel--drawer {
+  padding-bottom: 1.25rem;
+}
+
+@media (max-width: 767px) {
+  .token-request-kind-panel {
+    gap: 16px;
+    padding: 12px 14px 18px;
+  }
+
+  .token-request-kind-panel-header {
+    align-items: flex-start;
+  }
+
+  .token-request-kind-layout,
+  .token-request-kind-quick-filters,
+  .token-request-kind-columns {
+    grid-template-columns: minmax(0, 1fr);
+  }
+
+  .token-request-kind-clear {
+    min-height: 2.25rem;
+  }
+}
