+++ web/src/styles/user-console.css
@@ -1,167 +1,4 @@
 
-.dark .request-kind-badge {
-  --request-kind-ink: hsl(210 40% 92%);
-  --request-kind-border: hsl(217 20% 42% / 0.78);
-  --request-kind-bg-start: hsl(219 24% 24% / 0.96);
-  --request-kind-bg-end: hsl(219 22% 18% / 0.98);
-  --request-kind-chip: hsl(0 0% 100% / 0.12);
-  --request-kind-chip-border: hsl(0 0% 100% / 0.08);
-  box-shadow:
-    inset 0 1px 0 hsl(0 0% 100% / 0.08),
-    0 18px 34px -28px hsl(0 0% 0% / 0.7);
-}
-
-.dark .request-kind-badge--api,
-.dark .request-kind-badge--api-search {
-  --request-kind-ink: hsl(208 100% 84%);
-  --request-kind-border: hsl(210 94% 68% / 0.44);
-  --request-kind-bg-start: hsl(214 80% 20% / 0.96);
-  --request-kind-bg-end: hsl(210 72% 16% / 0.98);
-}
-
-.dark .request-kind-badge--api-extract {
-  --request-kind-ink: hsl(183 88% 79%);
-  --request-kind-border: hsl(184 78% 58% / 0.42);
-  --request-kind-bg-start: hsl(183 78% 18% / 0.96);
-  --request-kind-bg-end: hsl(186 68% 14% / 0.98);
-}
-
-.dark .request-kind-badge--api-crawl {
-  --request-kind-ink: hsl(42 100% 78%);
-  --request-kind-border: hsl(38 96% 60% / 0.44);
-  --request-kind-bg-start: hsl(32 76% 19% / 0.96);
-  --request-kind-bg-end: hsl(28 72% 15% / 0.98);
-}
-
-.dark .request-kind-badge--api-map {
-  --request-kind-ink: hsl(151 72% 78%);
-  --request-kind-border: hsl(151 62% 56% / 0.44);
-  --request-kind-bg-start: hsl(151 62% 18% / 0.96);
-  --request-kind-bg-end: hsl(153 56% 14% / 0.98);
-}
-
-.dark .request-kind-badge--api-research {
-  --request-kind-ink: hsl(334 92% 84%);
-  --request-kind-border: hsl(333 82% 66% / 0.42);
-  --request-kind-bg-start: hsl(330 66% 19% / 0.96);
-  --request-kind-bg-end: hsl(336 60% 15% / 0.98);
-}
-
-.dark .request-kind-badge--api-research-result {
-  --request-kind-ink: hsl(352 94% 84%);
-  --request-kind-border: hsl(352 84% 68% / 0.42);
-  --request-kind-bg-start: hsl(352 68% 20% / 0.96);
-  --request-kind-bg-end: hsl(355 62% 16% / 0.98);
-}
-
-.dark .request-kind-badge--api-usage {
-  --request-kind-ink: hsl(214 34% 83%);
-  --request-kind-border: hsl(216 28% 62% / 0.4);
-  --request-kind-bg-start: hsl(219 24% 21% / 0.96);
-  --request-kind-bg-end: hsl(220 20% 17% / 0.98);
-}
-
-.dark .request-kind-badge--api-raw {
-  --request-kind-ink: hsl(236 96% 85%);
-  --request-kind-border: hsl(236 80% 70% / 0.4);
-  --request-kind-bg-start: hsl(238 54% 21% / 0.96);
-  --request-kind-bg-end: hsl(239 48% 17% / 0.98);
-}
-
-.dark .request-kind-badge--mcp {
-  --request-kind-ink: hsl(204 100% 85%);
-  --request-kind-border: hsl(204 86% 68% / 0.42);
-  --request-kind-bg-start: hsl(208 52% 22% / 0.96);
-  --request-kind-bg-end: hsl(212 44% 18% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-search {
-  --request-kind-ink: hsl(191 100% 82%);
-  --request-kind-border: hsl(191 88% 62% / 0.44);
-  --request-kind-bg-start: hsl(192 74% 19% / 0.96);
-  --request-kind-bg-end: hsl(194 66% 15% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-extract {
-  --request-kind-ink: hsl(219 100% 86%);
-  --request-kind-border: hsl(220 84% 68% / 0.44);
-  --request-kind-bg-start: hsl(222 62% 21% / 0.96);
-  --request-kind-bg-end: hsl(226 54% 17% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-batch {
-  --request-kind-ink: hsl(25 100% 80%);
-  --request-kind-border: hsl(25 92% 63% / 0.42);
-  --request-kind-bg-start: hsl(24 72% 20% / 0.96);
-  --request-kind-bg-end: hsl(22 66% 16% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-tool {
-  --request-kind-ink: hsl(94 80% 80%);
-  --request-kind-border: hsl(94 64% 58% / 0.42);
-  --request-kind-bg-start: hsl(96 62% 18% / 0.96);
-  --request-kind-bg-end: hsl(100 56% 14% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-initialize {
-  --request-kind-ink: hsl(42 100% 82%);
-  --request-kind-border: hsl(40 92% 66% / 0.44);
-  --request-kind-bg-start: hsl(36 68% 20% / 0.96);
-  --request-kind-bg-end: hsl(32 62% 16% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-notification {
-  --request-kind-ink: hsl(320 96% 86%);
-  --request-kind-border: hsl(320 84% 68% / 0.44);
-  --request-kind-bg-start: hsl(318 58% 21% / 0.96);
-  --request-kind-bg-end: hsl(322 50% 17% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-ping {
-  --request-kind-ink: hsl(162 82% 82%);
-  --request-kind-border: hsl(162 68% 60% / 0.44);
-  --request-kind-bg-start: hsl(162 58% 18% / 0.96);
-  --request-kind-bg-end: hsl(166 50% 14% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-resource {
-  --request-kind-ink: hsl(278 100% 86%);
-  --request-kind-border: hsl(278 82% 70% / 0.44);
-  --request-kind-bg-start: hsl(276 52% 21% / 0.96);
-  --request-kind-bg-end: hsl(280 46% 17% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-tools-list {
-  --request-kind-ink: hsl(245 100% 87%);
-  --request-kind-border: hsl(244 82% 70% / 0.44);
-  --request-kind-bg-start: hsl(242 50% 21% / 0.96);
-  --request-kind-bg-end: hsl(246 44% 17% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-system {
-  --request-kind-ink: hsl(254 100% 86%);
-  --request-kind-border: hsl(252 86% 70% / 0.42);
-  --request-kind-bg-start: hsl(252 56% 21% / 0.96);
-  --request-kind-bg-end: hsl(255 48% 17% / 0.98);
-}
-
-.dark .request-kind-badge--mcp-raw {
-  --request-kind-ink: hsl(212 28% 84%);
-  --request-kind-border: hsl(214 22% 60% / 0.38);
-  --request-kind-bg-start: hsl(214 22% 21% / 0.96);
-  --request-kind-bg-end: hsl(216 18% 17% / 0.98);
-}
-
-.token-request-kind-menu .request-kind-badge {
-  width: auto;
-  max-width: none;
-}
-
-.token-request-kind-menu .request-kind-badge__detail {
-  overflow: visible;
-  text-overflow: clip;
-}
-
 /* Key/value information grid for compact readable summary */
 .kv-grid {
   display: grid;
